Add files via upload
The first potentiometer changes the color of the LEDs, i.e. by turning it we change one of the 10 selected colors. For example, in the position of the potentiometer from 0 degrees to 30 the color is white, from 31 to 60 red, from 61 to 90 green, etc.

The second potentiometer changes the speed at which the LEDs switch.

At any given time, only one LED is lit and after a while the next one lights up. The speed of change is regulated by potentiometer number 2. When the last LED is lit, the next one will be the first again.