Update second.ino
This code is a program for an Arduino board that controls a series of LEDs. The program declares three variables: zm1, zm2, and zm3, and initializes them to certain values. It then defines the setup() function, which is called only once when the program starts. Inside the setup() function, the program sets the mode of several pins to output and initializes the values of the zm1, zm2, and zm3 variables.

The main part of the program is the loop() function, which is called repeatedly in an infinite loop. Inside the loop() function, there is a for loop that iterates 10 times. Inside this loop, there is another for loop that iterates from 1 to 10. For each value of zm2, there is a third for loop that repeats a certain number of times, depending on the value of zm3.

Inside the third for loop, the program turns on an LED by setting the corresponding pin to HIGH, waits for a certain amount of time using the delay() function, turns off the LED by setting the pin to LOW, and then waits again. The delay time is determined by the value of the zm1 variable.

The program will turn on and off each LED in sequence, repeating the process a certain number of times depending on the value of zm3. The program will repeat this process 10 times in total.